Add validation for EnhancedUsersServiceImplementation

+++ src/main/java/dts/logic/EnhancedUsersServiceImplementation.java
@@ -13,6 +13,9 @@ import boundaries.UserBoundary;
 import dts.converter.UserConverter;
 import dts.dao.UserDao;
 import dts.data.UserEntity;
+import dts.data.UserRole;
+import exceptions.InvalidUserException;
+import exceptions.UserAlreadyExistsException;
 import exceptions.UserNotFoundException;
 import models.users.UserId;
 
@@ -33,6 +36,25 @@ public class EnhancedUsersServiceImplementation implements UsersService {
 	@Transactional
 	public UserBoundary createUser(UserBoundary user) {
 		UserEntity newUserEntity = this.userConverter.toEntity(user);
+		Optional<UserEntity> existingUser = this.userDao.findById(newUserEntity.getUserId());
+		if(existingUser.isPresent())
+		{
+			throw new UserAlreadyExistsException("User with email: " +
+					newUserEntity.getUserId() +	" and name " +
+					newUserEntity.getUsername() + " already exists");
+		}
+		if(newUserEntity.getUserId().isEmpty() || 
+		   newUserEntity.getUsername().isEmpty() ||
+		   newUserEntity.getUserId() == null ||
+		   newUserEntity.getUsername() == null ||
+		   !newUserEntity.getUserId().matches("^[A-Z0-9._%+-]+@[A-Z0-9.-]+\\.[A-Z]{2,6}$") ||
+		   !(newUserEntity.getRole().equals(UserRole.ADMIN) ||
+			 newUserEntity.getRole().equals(UserRole.MANAGER) ||
+			 newUserEntity.getRole().equals(UserRole.PLAYER)))
+		{
+			throw new InvalidUserException("Invalid email: " + newUserEntity.getUserId()
+					+ " or user name: " + newUserEntity.getUsername());
+		}
 		this.userDao.save(newUserEntity);
 		return this.userConverter.toBoundary(newUserEntity);
 	}
@@ -51,9 +73,18 @@ public class EnhancedUsersServiceImplementation implements UsersService {
 	public UserBoundary updateUser(UserBoundary update, String userSpace, String userEmail) {
 		Optional<UserEntity> existingUser = this.userDao.findById(new UserId(userSpace, userEmail).toString());
 		if (!existingUser.isPresent())
+		{			
 			throw new UserNotFoundException("user with email: " + userEmail + "does not exist");
+		}
 		UserEntity existingEntity = existingUser.get();
 		UserEntity userEntity = this.userConverter.toEntity(update);
+		if(!userEntity.getUserId().matches("^[A-Z0-9._%+-]+@[A-Z0-9.-]+\\.[A-Z]{2,6}$") ||
+				userEntity.getUserId() == null || 
+				userEntity.getUsername() == null)
+			{
+			throw new InvalidUserException("Invalid email: " + userEntity.getUserId()
+			+ " or user name: " + userEntity.getUsername());
+			}
 		userEntity.setUserId(existingEntity.getUserId().toString());
 		return this.userConverter.toBoundary(this.userDao.save(userEntity));
 	}
@@ -70,6 +101,18 @@ public class EnhancedUsersServiceImplementation implements UsersService {
 	@Override
 	@Transactional
 	public void deleteAllUsers(String adminSpace, String adminEmail) {
+		Optional<UserEntity> existingAdmin = this.userDao.findById(new UserId(adminSpace, adminEmail).toString());
+		if(!existingAdmin.isPresent())
+		{
+			throw new UserNotFoundException("user with email: " + existingAdmin + "does not exist");
+		}
+		UserEntity existingAdminEntity = existingAdmin.get();
+		if(!existingAdminEntity.getRole().equals(UserRole.ADMIN))
+		{
+			throw new InvalidUserException("Invalid email: " + existingAdminEntity.getUserId()
+			+ " or user name: " + existingAdminEntity.getUsername());
+		}
+
 		this.userDao.deleteAll();
 	}
